Merge remote-tracking branch 'origin/master'
# Conflicts:
#	ActivityLauncherApp/src/main/java/com/kapcode/parentalcontrols/ParentalControlService.java

changed the way we bring user to home screen. instead of launching the launcher by finding the user highlighted launcher, we just use ParentalControlService.goHome() which launches androids intent to the default launcher activity.

It is discovered that this will not work on newer android builds. neither will the old way, the back end is essentialy the same thing (launching the intent of the launchers activity).  Further permission is required to launch intents from background processes on newer versions of android.

Everything works fine on targeted amazon tablets.

fixed